+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer.Test/Helpers/CodeFixVerifier.Helper.cs
