+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer.Test/Helpers/DiagnosticVerifier.Helper.cs
