+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer.Test/MindTouchMaterializeCollectionsAnalyzer.Test.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{8B713886-9740-453A-B438-7EC8F68889DA}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>MaterializeCollectionsAnalyzer.Test</RootNamespace>
-    <AssemblyName>MaterializeCollectionsAnalyzer.Test</AssemblyName>
+    <RootNamespace>MindTouchMaterializeCollectionsAnalyzer.Test</RootNamespace>
+    <AssemblyName>MindTouchMaterializeCollectionsAnalyzer.Test</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
@@ -114,13 +114,13 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\MaterializeCollectionsAnalyzer\MaterializeCollectionsAnalyzer.csproj">
-      <Project>{CEFDD049-E7DE-4503-AC24-2615EB6B1148}</Project>
-      <Name>MaterializeCollectionsAnalyzer</Name>
-    </ProjectReference>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <ItemGroup>
-    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
+    <ProjectReference Include="..\MindTouchMaterializeCollectionsAnalyzer\MindTouchMaterializeCollectionsAnalyzer.csproj">
+      <Project>{cefdd049-e7de-4503-ac24-2615eb6b1148}</Project>
+      <Name>MindTouchMaterializeCollectionsAnalyzer</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
