+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer.Vsix/source.extension.vsixmanifest
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <PackageManifest Version="2.0.0" xmlns="http://schemas.microsoft.com/developer/vsx-schema/2011" xmlns:d="http://schemas.microsoft.com/developer/vsx-schema-design/2011">
   <Metadata>
-    <Identity Id="MaterializeCollectionsAnalyzer..294abe5b-1773-41cc-b48c-277053a1a713" Version="1.0" Language="en-US" Publisher="Admin"/>
-    <DisplayName>MaterializeCollectionsAnalyzer</DisplayName>
+    <Identity Id="MindTouchMaterializeCollectionsAnalyzer..294abe5b-1773-41cc-b48c-277053a1a713" Version="1.0" Language="en-US" Publisher="Admin"/>
+    <DisplayName>MindTouchMaterializeCollectionsAnalyzer</DisplayName>
     <Description xml:space="preserve">Roslyn analyzer that ensures that any collections passed in as arguments or returned from a function are materilized.Roslu</Description>
   </Metadata>
   <Installation>
@@ -12,7 +12,7 @@
     <Dependency Id="Microsoft.Framework.NDP" DisplayName="Microsoft .NET Framework" d:Source="Manual" Version="[4.5,)" />
   </Dependencies>
   <Assets>
-    <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="MaterializeCollectionsAnalyzer" Path="|MaterializeCollectionsAnalyzer|"/>
-    <Asset Type="Microsoft.VisualStudio.Analyzer" d:Source="Project" d:ProjectName="MaterializeCollectionsAnalyzer" Path="|MaterializeCollectionsAnalyzer|"/>
+    <Asset Type="Microsoft.VisualStudio.MefComponent" d:Source="Project" d:ProjectName="MindTouchMaterializeCollectionsAnalyzer" Path="|MindTouchMaterializeCollectionsAnalyzer|" />
+    <Asset Type="Microsoft.VisualStudio.Analyzer" d:Source="Project" d:ProjectName="MindTouchMaterializeCollectionsAnalyzer" Path="|MindTouchMaterializeCollectionsAnalyzer|" />
   </Assets>
 </PackageManifest>
