+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer/CodeFixProvider.cs
@@ -21,12 +21,12 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CodeActions;
+using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace MaterializeCollectionsAnalyzer {
+namespace MindTouchMaterializeCollectionsAnalyzer {
 
     [ExportCodeFixProvider(LanguageNames.CSharp, Name = nameof(MaterializeCollectionsAnalyzerCodeFixProvider)), Shared]
     public class MaterializeCollectionsAnalyzerCodeFixProvider : CodeFixProvider {
