+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer/MaterializedCollectionsUtils.cs
@@ -20,7 +20,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 
-namespace MaterializeCollectionsAnalyzer {
+namespace MindTouchMaterializeCollectionsAnalyzer {
     internal class MaterializedCollectionsUtils {
 
         //--- Class Methods ---
