+++ MindTouchMaterializeCollectionsAnalyzer/MindTouchMaterializeCollectionsAnalyzer/Resources.resx
